backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf_yasg==1.21.9
python-dotenv==1.0.1
rest_framework_simplejwt==0.0.2
psycopg2-binary==2.9.5
gunicorn==20.1.0
django-filter==23.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.14.0
whitenoise==6.5.0
Pillow==9.5.0

